wip(go): delete access tokens only if tenant scoped;

+++ logic/auth.go
@@ -378,6 +378,10 @@ func DeleteUser(ctx context.Context, user *schema.User) error {
 	}
 
 	RemoveUserFromAclPolicy(ctx, user.Username)
+
+	if scope.Level(ctx) != scope.TenantScope {
+		return nil
+	}
 	return (&schema.UserAccessToken{UserName: user.Username}).DeleteAllUserTokens(ctx)
 }
